Store boundary residuals as float32: gemma's non-finite turn-start readouts were an fp16 overflow
make_projection_hook stored the event-position residual as float16. gemma-3-12b-it's
prefill row carries one coordinate above the float16 max (65,504), so that single
coordinate became inf and the whole vector read as non-finite. Across all 629
non-finite boundary residuals in seven gemma conditions, exactly one of 3,840
coordinates had overflowed every time. The per-token proj/norm arrays were never
affected -- they are reduced in float32 before any cast.

Storing float32 costs 20 KB twice a turn. The float16 saving in the comment above
applies to the directions payload, not to this. The dashboard's Generation.arrays
downcast the same two vectors on save and is fixed with it.

Records written before this change keep their float16 residuals and the skip
counts in docs/measurement.md stand for them.

Also corrects the Ministral affpos6 depth figures in docs/runs.md: at the now
complete n=24 that cell reaches t5 in 7 of 24 rollouts, not 20, because the affect
prompt lifts its solve rate to 16/24 and 15 rollouts are single-turn. aff6 in the
same table is 24, not 23. Both verified against the records, counting non-empty
turns per the documented turn-indexing rule.

+++ src/healthy_rl/dashboard/generation.py
@@ -121,10 +121,13 @@ class Generation:
         out = {"proj": self.proj.astype(np.float32), "norm": self.norm.astype(np.float32),
                "proj_prefill": self.proj_prefill.astype(np.float32),
                "norm_prefill": self.norm_prefill.astype(np.float32)}
+        # float32 for the same reason the hook stores them that way: a prefill
+        # row can hold a coordinate above the float16 max and the cast makes the
+        # whole vector non-finite (docs/measurement.md, "Non-finite residuals").
         if self.res_start is not None:
-            out[f"res_start_L{probe_layer}"] = self.res_start.astype(np.float16)
+            out[f"res_start_L{probe_layer}"] = self.res_start.astype(np.float32)
         if self.res_end is not None:
-            out[f"res_end_L{probe_layer}"] = self.res_end.astype(np.float16)
+            out[f"res_end_L{probe_layer}"] = self.res_end.astype(np.float32)
         return out
